Fix: Fixed a bug where startuptaints weren't correctly injected into bottlerocket (#2327)
