+++ pkg/cloudprovider/aws/amifamily/resolver.go
@@ -107,8 +107,15 @@ func (r Resolver) Resolve(ctx context.Context, provider *v1alpha1.AWS, nodeReque
 	var resolvedTemplates []*LaunchTemplate
 	for amiID, instanceTypes := range amiIDs {
 		resolved := &LaunchTemplate{
-			Options:             options,
-			UserData:            amiFamily.UserData(nodeRequest.Template.KubeletConfiguration, nodeRequest.Template.Taints, options.Labels, options.CABundle, instanceTypes, aws.String(userDataString)),
+			Options: options,
+			UserData: amiFamily.UserData(
+				nodeRequest.Template.KubeletConfiguration,
+				append(nodeRequest.Template.Taints, nodeRequest.Template.StartupTaints...),
+				options.Labels,
+				options.CABundle,
+				instanceTypes,
+				aws.String(userDataString),
+			),
 			BlockDeviceMappings: provider.BlockDeviceMappings,
 			MetadataOptions:     provider.MetadataOptions,
 			AMIID:               amiID,
