+++ pkg/cloudprovider/aws/testdata/br_userdata_unmerged.golden
@@ -10,3 +10,5 @@ max-pods = 110
 'karpenter.sh/provisioner-name' = '%s'
 
 [settings.kubernetes.node-taints]
+baz = ['bin:NoExecute']
+foo = ['bar:NoExecute']
