+++ test/suites/integration/testdata/br_userdata_input.golden
@@ -1,2 +1,4 @@
 [settings.kubernetes]
 kube-api-qps = 30
+[settings.kubernetes.node-taints]
+"node.cilium.io/agent-not-ready" = ["true:NoExecute"]
